Change mobile layout of tabs:
On smaller screens tabs are now inline, but only contain titles. When a
tab is active it populates an area below with trend info. Only trend info
related to the currently active tab can be seen on smaller devices. This
is a tradeoff between vertical space and showing all possible
information at all times.