Fix chart heading
After the refactoring of the chart AJAX the helper was no longer
possible as the request was not being routed through the Audience
controller.

Appended the title to the JSON response and load the title through JS,
see the `updateChartHeader` function in `areas.coffee`.

It's a little unfortunate we have to load the title via JS but this is
only textual content we need so still acceptable.